pyproject.toml
[tool.poetry]
name = "slimconvert"
version = "0.1.0"
package-mode = false
description = ""
authors = ["Vladislav <113159483+Ada-lave@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
python-docx = "^1.1.2"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
